requirements.txt
attrs==21.4.0
click==8.0.4
dataclasses==0.8
Flask==2.0.3
Flask-Cors==3.0.10
importlib-metadata==4.8.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pkg-resources==0.0.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
python-dotenv==0.19.2
six==1.16.0
tomli==1.2.3
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
